move osAiSetNextBuffer
