--- src/audio/internal/dcache.c
@@ -1,16 +1,0 @@
-#include "ultra64.h"
-#include "z64audio.h"
-
-void Audio_InvalDCache(void* buf, s32 size) {
-    OSIntMask prevMask = osSetIntMask(OS_IM_NONE);
-
-    osInvalDCache(buf, size);
-    osSetIntMask(prevMask);
-}
-
-void Audio_WritebackDCache(void* buf, s32 size) {
-    OSIntMask prevMask = osSetIntMask(OS_IM_NONE);
-
-    osWritebackDCache(buf, size);
-    osSetIntMask(prevMask);
-}
