+++ src/audio/internal/os.c
@@ -1,4 +1,23 @@
+/**
+ * Original Filename: os.c
+ */
+
 #include "ultra64.h"
+#include "z64audio.h"
+
+void Audio_InvalDCache(void* buf, s32 size) {
+    OSIntMask prevMask = osSetIntMask(OS_IM_NONE);
+
+    osInvalDCache(buf, size);
+    osSetIntMask(prevMask);
+}
+
+void Audio_WritebackDCache(void* buf, s32 size) {
+    OSIntMask prevMask = osSetIntMask(OS_IM_NONE);
+
+    osWritebackDCache(buf, size);
+    osSetIntMask(prevMask);
+}
 
 /**
  * Submits an audio buffer to be consumed by the Audio DAC. The audio interface can queue a second DMA while another
